Move ServoSelectorImpl to a dedicated module.
