+++ components/layout_thread/lib.rs
@@ -107,8 +107,8 @@ use style::media_queries::{Device, MediaType};
 use style::parallel::WorkQueueData;
 use style::properties::ComputedValues;
 use style::refcell::RefCell;
-use style::selector_matching::USER_OR_USER_AGENT_STYLESHEETS;
 use style::servo::{Animation, LocalStyleContextCreationInfo, SharedStyleContext, Stylesheet, Stylist};
+use style::servo_selector_impl::USER_OR_USER_AGENT_STYLESHEETS;
 use style::stylesheets::CSSRuleIteratorExt;
 use style::workqueue::WorkQueue;
 use url::Url;
